refactor: useThemeInfiniteScroll 훅 레벨에 따른 추상화
