+++ src/page/Themes/hooks/useParamsIndex.ts
@@ -1,0 +1,7 @@
+import { useParams } from "react-router-dom";
+
+export const useParamsIndex = () => {
+  const { id } = useParams<{ id: string }>();
+  const index = Number(id);
+  return index;
+};
